# spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  title: Gamification API
  description: Move your app forward with the RFPS API.
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
paths:
...
  /pointScales:
    get:
      description: 
        The PointScales endpoint returns information about point scales which are available. The response includes the id, the name, the description, and the coefficient of each point scale. This response is a list of point scale sorted by the id.
      produces:
        - application/json
      tags:
        - PointScales
      responses:
        200:
          description: OK, an array of point scales.
          schema:
            type: array
            items:
              $ref: '#/definitions/PointScaleOutputDTO'
        500:
          description: If database isn't reachable.
          
    post:
      description: Create a new point scale.
      consumes:
        - application/json
      parameters:
        - in: body
          name: badge
          description: PointScale object that needs to be added to the database.
          required: true
          schema:
              $ref: '#/definitions/PointScaleInputDTO'
        - name: token
          in: header
          type: integer
          format: int64
          description: The application for which the badge belong
          required: true
      responses:
        201:
          description: Created, returns the URL to access the new point scale.
          schema:
            $ref: '#/definitions/LocationPointScale'
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if mandatory fields contain more than 80 characters,
                       if point scale name provided already exists.
        500:
          description: If database isn't reachable.
  /pointScales/{id}:
    get:
      description: Returns the point scale which has the specified id.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the desired point scale.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/PointScaleOutputDTO'
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
          
    put:
      description: Modify completely the existing point scale which has the specified id.
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the point scale to modify.
          required: true
          type: string
        - name: pointScale
          in: body
          description: PointScale to be updated
          required: true
          schema:
              $ref: '#/definitions/PointScaleInputDTO'
              
      responses:
        204:
          description: No content, if success.
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if pointScale name provided already exists.
          

    delete:
      description: Delete the point scale which has the specified id.
      parameters:
        - name: id
          in: path
          description: The id of the desired point scale.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        404:
          description: Not found, if id provided don't exist.
        500:
          description: If database isn't reachable.
...
definitions:
...
  PointScaleOutputDTO:
    type: object
    required: [name,description, coefficient]
    properties:
      pointScale_id:
        type: string
        description: Unique identifier representing a specific point scale.
      name:
        type: string
        description: Display name of point scale.
      description:
        type: string
        description: Description of point scale.
      coefficient:
        type: integer
        description: Coeffcient of the point scale.
  PointScaleInputDTO:
    type: object
    required: [name,description, coefficient]
    properties:
      name:
        type: string
        description: Display name of point scale.
      description:
        type: string
        description: Description of point scale.
      coefficient:
        type: integer
        description: Coeffcient of the point scale.
  LocationPointScale:
    type: string 
    required: [locationPointScale]
    properties:
      locationPointScale:
        type: string
        description: Unique URL to access the new object created.
...